saglik-petegim/apps/backend_app/requirements.txt
# Production Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
boto3==1.34.0
pillow==10.1.0
jinja2==3.1.2
aiofiles==23.2.0
httpx==0.25.2
cryptography==41.0.8
slowapi==0.1.9
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
apscheduler==3.10.4

# Supabase Integration
supabase==2.3.0
postgrest==0.13.0
gotrue==1.3.0
realtime==1.0.0
storage3==0.6.0

# RAG and AI Dependencies
pinecone-client[grpc]==3.0.2
langchain-pinecone==0.0.3
langchain-google-genai==1.0.1
langchain==0.1.6
langchain-community==0.0.19
pypdf==4.0.1
tiktoken==0.5.2

# Development Dependencies (for local development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-benchmark==4.0.0
pytest-timeout==2.2.0
pytest-html==4.1.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
coverage==7.3.2
pytest-cov==4.1.0
factory-boy==3.3.0
locust==2.17.0
faker==20.1.0
freezegun==1.2.2
respx==0.20.2
aioresponses==0.7.4
pytest-postgresql==5.0.0
testcontainers==3.7.1